Ensure case order quantities are entered as whole numbers
Update the purchase order detail page to use `parseInt` for case quantity input, preventing fractional case orders.

+++ client/src/pages/purchase-order-detail.tsx
@@ -538,7 +538,7 @@ export default function PurchaseOrderDetail() {
                                   step="1"
                                   min="0"
                                   value={caseQty || ""}
-                                  onChange={(e) => handleCaseQuantityChange(itemId, parseFloat(e.target.value) || 0)}
+                                  onChange={(e) => handleCaseQuantityChange(itemId, parseInt(e.target.value) || 0)}
                                   onKeyDown={(e) => handleKeyDown(e, itemId, displayItems)}
                                   className="text-center"
                                   placeholder="0"
